requirements.txt
fastapi
uvicorn
pyjwt
passlib
python-multipart
bcrypt
peewee
sqlalchemy